requirements.txt
attrs==25.4.0
Automat==25.4.16
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.20.0
Flask==3.1.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.12.2
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.2
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
packaging==25.0
parsel==1.10.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
pyOpenSSL==25.3.0
python-dotenv==1.2.1
queuelib==1.8.0
requests==2.32.5
requests-file==3.0.1
Scrapy==2.13.4
service-identity==24.2.0
tldextract==5.3.0
Twisted==25.5.0
typing_extensions==4.15.0
urllib3==2.5.0
w3lib==2.3.1
Werkzeug==3.1.4
wheel==0.46.1
zope.interface==8.1.1
